requirements.txt
requests~=2.32.5
pandas~=2.3.3
python-dateutil~=2.9.0.post0
selectolax~=0.4.0
selenium~=4.38.0